fix: GetWorldStockmonsRequestDto - latitude 타입 수정

+++ core/src/main/java/com/pda/core/dto/GetWorldStockmonsRequestDto.java
@@ -13,7 +13,7 @@ import static com.pda.core.exception.ExceptionMessage.OUT_OF_WORLD_RANGE;
 public class GetWorldStockmonsRequestDto {
     @DecimalMin(value = "37.42928001105901", message = OUT_OF_WORLD_RANGE)
     @DecimalMax(value = "37.73892900008836", message = OUT_OF_WORLD_RANGE)
-    private final int latitude;
+    private final double latitude;
 
     @DecimalMax(value = "127.21925741327027", message =  OUT_OF_WORLD_RANGE)
     @DecimalMin(value = "126.76605328501563", message =  OUT_OF_WORLD_RANGE)
